+++ gemrb/plugins/Core/GameScript.cpp
@@ -15,7 +15,7 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA 02111-1307, USA.
  *
- * $Header: /data/gemrb/cvs2svn/gemrb/gemrb/gemrb/plugins/Core/GameScript.cpp,v 1.373 2006/08/09 19:04:33 avenger_teambg Exp $
+ * $Header: /data/gemrb/cvs2svn/gemrb/gemrb/gemrb/plugins/Core/GameScript.cpp,v 1.374 2006/08/11 23:17:19 avenger_teambg Exp $
  *
  */
 
@@ -1098,7 +1098,7 @@ GameScript::GameScript(ieResRef ResRef, unsigned char ScriptType,
 		freeLocals = true;
 	}
 	scriptlevel = ScriptLevel;
-	lastAction = ~0;
+	lastAction = (unsigned int) ~0;
 
 	if (!initialized) {
 		initialized = 1;
