+++ gemrb/plugins/Core/VFS.cpp
@@ -15,7 +15,7 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
  *
- * $Header: /data/gemrb/cvs2svn/gemrb/gemrb/gemrb/plugins/Core/VFS.cpp,v 1.17 2006/01/14 17:16:42 avenger_teambg Exp $
+ * $Header: /data/gemrb/cvs2svn/gemrb/gemrb/gemrb/plugins/Core/VFS.cpp,v 1.18 2006/08/11 23:17:20 avenger_teambg Exp $
  *
  */
 
@@ -280,7 +280,7 @@ char* PathJoin (char* target, ...)
  */
 void FixPath (char *path, bool needslash)
 {
-	unsigned int i = strlen( path ) - 1;
+	size_t i = strlen( path ) - 1;
 
 	if (needslash) {
 		if (path[i] == PathDelimiter) return;
